Test and fix the proof that all naturals are at least 0.
Some bugs were discovered in the primitives.  There was a bug in
logic.linear, but it was a trivila careless mistake regarding the
curry_howard correspondents of linear OnBody arrows.

The proof that all naturals are at least 0 can now be accurately
translated into a working program.  Note that in order to understand its
output you should realize that we're currently using a quirky
representation of a <= b for illustrative purposes.  a <= b is
represented as a list of successive integers in the closed-closed
interval (a, b).  While the null representation is also appropriate,
this representation better showcases the way primitives are built into
programs by the curry_howard correspondents of proofs.