Organizational commit.  Change the names of things.
Create a new directory structure to house the code.
Move files into the new directory structure, giving them better names.

Change the names of certain variables and functions to make them
clearer.  Most notably, we no longer refer to a "linear" calculus and a
"linearui" calculus.  Instead, we call them the basic calculus and the
enriched calculus repsectively.

Libraries to use as starting points for proofs are collected in "lib".
Example proofs are collected in "examples".

Simplify the translations between different representations of proof:
  calculus.basic defines the category of basic formulas and proofs.
  calculus.enriched defines the category of enriched formulas and
    proofs.
  extraction contains a directory for each of fantasia's embedded
    extraction engines. Currently the only extraction engine is the
    python extraction engine which represents formulas as certain python
    objects.
  each extraction E engine contains an arrows module which implements
    curry_howard on basic arrows.  E also contais "lib" and "examples"
    directories.  There directories should contain representations of
    the objects and arrows defined in the top level "lib" and "examples"
    directories.