Implement forwardHeavyClean transitions and simplify easy_induction.
In the spirit of forwardClean, add forwardHeavyClean to the enriched
calculus.  Now combinators can optionally implement more serious
simplifications.  These simplifications are intended to produce a new
formula that:
  a) follows logically from the src formula
  b) lacks much of the "unnecessary cruft" that may have existed in the
     src formula.

Each combinator may use its own definition of unnecessary cruft.  The
choice of what counts as cruft is really a UI decision and could be made
in multiple ways.  Combinators should pick definitions of cruft that
accord well with things the user probably never wants to have to look at
in his formulas.

It would be nice but not necessary for the results of calls to
forwardHeavyClean to be easily predictable.  As long as the resulting
formula is what the user wants to see, it doesn't matter how
elaborate an arrow it to get there.

examples/easy_induction now makes use of forwardHeavyClean to further
simplify its proof.