Add a way to generate *transpose* methods automatically.
The design decision in this commit is to complicate the description of
how transpose, transposeToNot, and notToTranspose work.  The complexity
currently isn't terrible, but it is ugly.

The value in introducing this complexity is that it allows the code to
simultaneously satisfy the constraints of:
  a) when implementing new primitive, there should be no need to define
  two seperate primitive classes each of which is the transpose of the
  other.
    [ This constraint isn't necessary, but it seems desirable, and it
      saves (?a lot?) of work. ]

  b) for every enriched.Logic x, x.transpose().transpose() == x
    [ This constraint currently appears absolutely necessary.  Firstly,
      it makes a tremendous amount of sense.  Secondly, many of the
      implementation of notToTranspose and transposeToNot are
      irreparably dependent on this constraint. ]

The obvious alternative to introducing this complexity is to drop
constraint a).  In this case, for every new primitive enriched.Logic,
we would need to make sure there are two kinds of this primitive, the
positive kind p and the negative kind n.  Also, we would need to define
new formulas in a library for converting Not(p) <--> n and Not(n) <--> p

Maybe someone will develp a prettier solution than this one later on.
For now, let's hope it works well.