Correct notToTranspose and transposeToNot. Test definitions.
There were several problems with the implementations of notToTranspose
and transposeToNot.  These have been corrected enough to create valid
transitions that operate on a simple example definition transition.

A convenient consequence of this commit is that we can now translate
enirched definition arrows into the basic calculus and then convert them
into programs.