Implement a new kind of relation holds primitive.
There were several problems with the previous treatment of primitives.

  0) There appears to be no need to distinguish between the basic and
     the enriched primitives.
  1) The old approach involved defining a new class every time you
     wanted to deal with a new primitive.  This requirement means that
     writing code for new primitives is harder, and it confounded the
     user's ability to create primitives within fantasia.
  2) Primitives were defined in the calculus.basic and calculus.enriched
     modules, thereby occluding the moandic nature of the calculi.
     The calculi themselves do not define a category, rather they define
     a way in which a category can be extended.

The new relation primitives are rendered much more nicely than the
previous ones, and do not suffer the above limitations.
Also, they are designed to simplify Holds so that it is always binary.
These relation primitives have the further advantage of allowing for a
very elegant approach to defining an appropriate notion of equality and
functions.  Further details will follow in a later commit.