Factor constructors.py, continue the example, enhance libraries.
The Library and Proof classes now make better use of the structure of
the formulas they generate.  In particular, each formula is known to
start with an exists and be followed by a list of the library claims.
The default Proof objects now start with convenient arrow for beginning
a proof.  Also, the proof objects have a forwardFollow method that
applies its argument function at the appropriate spot inside the
formula.

constructors.py has now been factored into a basicConstructors file for
constructors that only depend on the basic calculus, along with a new
constructors file that contains all the basic constructors, along with
constructors for the constructs employed by the enriched calculus.
Certain fundamental modules (like equivalence) will import the basic
constructors module, but new modules should attempt to import only from
the enriched constructors module.