Implement an enriched holds combinator and test.
The enriched holds combinators allows the use of the Apply class when
referring to variables.  Translation of this construct is intricate
because it must separate the syntax for variables with that for objects.
For example, if an Apply occurs deep within a complex variable,
translation of the whole variable must draw certain objects generated
from the Apply translation outside of the complex variable.  To this
end, the translation code makes heavy use of CPS.