Implement compression of basic arrows.
There's a bit of trickery in preventing compression from taking
exponential time.  This commit takes great pains to avoid exponential
blowup.

Compression can still be improved.  Currently, compression exploits the
associative nature of composition, that Id is a unit for composition,
that And and Or are symmetric, and the F(a) o F(b) = F(a o b) law for
functors.  There are a great deal of other properties to exploit for
building better compression.  Leave a TODO to implement them later.