Start to unenrich the calculus.
This commit marks the start of an effort to remove all enriched
constructs from fantasia.  This involves deleting calculus.enriched and
modifying all uses of enriched constructs to use basic constructs
instead.  Consequently there will be no need for translate() methods.

The primary advantage of using the enriched calculus was that it
produced a substantially more compact notation.  That work will now have
to be done by rendering.  The basic calculus may therefore want to
contain hints as to how things should be rendered.  For example, it can
hint that a variable y should be rendered as the image f(x) and it can
hint that certain relations (e.g. "f(x) is natural") need not be
rendered at all.

The enriched calculus is being removed because it dramatically
complicates the metaprogramming and the implementation of certain
arrows.  Specifically,
  a) Instantiating a variable in a Forall is TREMENDOUSLY complex when
  the variable can contain an Apply object.
  b) The implementation of import and export has more cases when the
  calculus contains a BoundedExists construct.  Also, the import and
  export arrows all need to be translated.
  c) Translation of the EnrichedHolds construct is complex, and produces
  redundant formulas.