Correct python closure conversion bugs.
If you create an inner python function inside a loop or a list
comprehension, and the function is called after the loop increments, the
closure converted variables will be bound to the values of the new
iteration of the loop, not the old one.

This behavior is the most hideous and unconscionable property of the
python language.  An effective, yet bothersome hack is to recapture all
such variables as named argument parameters to the lambda.  This
technique is necessary in all cases in fantasia in which a function
returns a list, each of whose members contains a natural transform.  Fix
those inner functions.