Implement simplify and invert.  Improve rendering of functions.
This commit introduces a great deal of complexity into the proess of
rendering.  Rendering now intelligently replaces all occurences of a
variable B known to be the image of a variable A under a function f with
a special a-->f construct.  The replacements are recursive.  Also,
rendering removes superfluous claims stating that b in is the domain of
f.

These enhancements to rendering greatly reduce the size of many formulas
without impacting the clarity of what they express.

This kind of rendering dangerously divorces the user's understanding of
his formula from the actual formula.
The hope is that this divorce will not be a problem since interfaces
will still be able to do the things the user expects.  Future work will
elucidate whether or not this hope is justified.