Redefine the WellDefined functor using bifunctors.
Remove the WellDefined class and replace it with a complex composite of
endofunctors and bifunctors.

Illustrate the power and flexibility of bifunctors.  A single complex
class is now obviated by certain simple functors combined in simple
ways.

Ultimately, this type of complex bifunctor magic will be used to
translate enriched functors.  The most pressing example is the enriched
WellDefined endofunctor, but others may also arise.  For example, a
functor which expresses that something is well defined up to many
equivalence relations, or up to some custom formula.