Implement transport_duplicating methods. Remove enriched transport.
Enriched transport does not translate nicely into calls to basic
transport.  Define a new method, transport_duplicating.  Implement this
method in basic.  Implement this method in enriced via the basic method.
transport itself is not so necessary for the ui, and
transport_duplicating should a) suffice, b) be easier to use.