Implement enriched search.
To run an enriched search, you must define a SearchSpec instance which
somehow describes the enriched claim that you're searching for.  Pass
the SearchSpec instance to the search() method of an enriched
endofunctor, and it will compute a list of all enriched claims which it
can import/export provided it is covariant/contravariant respectively.

Later commits will define ways to actually do the imports via
translation to basic endofunctors.

Write some simple tests for searches through And and Composite
endofunctors.