Implement search and import on paths.  Write and pass some tests.
paths now have a search method that takes a SearchSpec instance and
computes both the importable claims matching the SearchSpec instance and
thunked arrows that actually do the imports.

Warning, the current implementation takes few pains to insure reasonable
performance.  Searching is essentially done twice. Searching is done
over the enriched formula to find which claims can be imported.  After
that, a second search is performed over the translate of the enriched
formula in order to locate the translate of the formula to be imported.
In principle, it should be possible to use the information found by the
first search about the location of the imported claim to completely
eliminate the need for the second search.  Doing so, however, would
require writing substantially more code (possibly even inventing new
bifunctors corresponding to the assumptions made by enriched Exists
endofunctors) and is therefore avoided in the current implementation.
Furthermore, there exists a decent reason to believe that the double
search will not cause performance problems:
  The searches will take time roughly linear in the depth of the
  searched endofunctor (not in its size).