Implement DependentArrows, define import, and begin substitution.
DependentArrows are like basic arrows, but need additional assumptions
in order to operate.  The assumptions made by dependent arrows will
typically be imported from a surrounding endofunctor.

Certain arrows (specifically the variable substitution arrows) should be
easier to generate by means of the DependentArrow/import paradigm.

Begin writing the code for variable substitution.

Note: a method already exists to substitute for the variables in an
formula, but this method does not create an arrow and therefore can not
be used in proofs.  The code for creating an arrow for variable
substitution is much more sophisticated and is computationally necessary
(i.e. code critical to a properly functioning algorithm will be
extracted from variable substitution arrows).