Implement more imports and enriched arrows and continue the qr proof.
The current work on the qr proof makes lots of compromises.  They are
made so that the qr proof can be finished, but future work should
refactor much of the
import/export/instantiate/importAbout/importAboutNegating/simplify
and Instantiator/SearchSpec code so as to make such macroscopic proof
operations easier to implement, simpler to understand, more covenient to
use, and more efficient to run.

Performance is starting to become a problem in the qr proof. As a
temporary workaround, modify the current QR proof to easily compress and
cache work in a file and load it later very quickly.