Document the interfaces to the proof midlevel.
Do not implement any non-trivial methods.
Define and comment methods of proofs and extenders to be implemented
later.  Attempt to give a complete description of the behavior of the
midlevel.

The interfaces defined in this commit are a rough draft.  Later commits
should create tests using these interfaces to test how complete they
are.  The interfaces may then be revised, and then implemented.

The interfaces to the proof midlevel divide the functionality of the
midlevel into about 6 separate loosely coupled classes called Extenders.
Each extender is responsible for generating a specific kind of proof
extension.

Expect that the implementation of extenders will be far more complex
than the interfaces to them.  For example, while the interface makes the
extenders appear as though they have little to do with one another, the
implementation of any one extender will probably use the other extenders
in complex ways.