Rewrite the factorization interface to use SearchResults.
The factorization interface now defines a factorization more simply.  A
factorization is simply an endofunctor F a bifunctor B, an endofunctor G
and a formula X such that:
  F(.) == B(., G(X))

The factor function in calculus.enriched.factor no longer returns
factorizations.  Instead it returns SearchResults.  Each search result
is essentially a pair of factorizations, one is the original
factorization, the other is the result of instantiating that
factorization with the requested variables.